refactor: simplify code-example-evaluator route model loading
Remove loading and filtering of evaluation subsets (pass, fail, unsure,
trusted) and drop helper methods for building filters and loading evaluations.
Load all languages instead of filtered languages. Streamline the model method
to simplify data requirements and reduce complexity. This change improves
maintainability by removing unused filtering logic and unused related types.
