+++ app/controllers/course-admin/code-example-evaluator.ts
@@ -3,7 +3,10 @@ import { action } from '@ember/object';
 import type RouterService from '@ember/routing/router-service';
 import { service } from '@ember/service';
 import type Store from '@ember-data/store';
+import { tracked } from '@glimmer/tracking';
 import type CommunitySolutionEvaluationModel from 'codecrafters-frontend/models/community-solution-evaluation';
+import type CommunitySolutionEvaluatorModel from 'codecrafters-frontend/models/community-solution-evaluator';
+import type CourseModel from 'codecrafters-frontend/models/course';
 import type CourseStageModel from 'codecrafters-frontend/models/course-stage';
 import type LanguageModel from 'codecrafters-frontend/models/language';
 import type { CodeExampleEvaluatorRouteModel } from 'codecrafters-frontend/routes/course-admin/code-example-evaluator';
@@ -21,6 +24,10 @@ export default class CodeExampleEvaluatorController extends Controller {
   usernames = '';
   course_stage_slugs = '';
 
+  @tracked passEvaluations: CommunitySolutionEvaluationModel[] = [];
+  @tracked failEvaluations: CommunitySolutionEvaluationModel[] = [];
+  @tracked unsureEvaluations: CommunitySolutionEvaluationModel[] = [];
+
   get currentCourseStage() {
     return this.filteredCourseStages[0] || null;
   }
@@ -34,11 +41,88 @@ export default class CodeExampleEvaluatorController extends Controller {
   }
 
   get filteredLanguages() {
-    return this.model.languages.filter((language) => this.model.filteredLanguageSlugs.includes(language.slug));
+    return this.model.allLanguages.filter((language) => this.model.filteredLanguageSlugs.includes(language.slug));
+  }
+
+  get isLoadingEvaluations() {
+    return this.loadEvaluationsTask.isRunning;
   }
 
   get sortedLanguagesForDropdown() {
     return this.model.course.betaOrLiveLanguages.toSorted(fieldComparator('name'));
+  }
+
+  buildContextFilters(_course: CourseModel, languageSlugsFilter: string[], courseStageSlugsFilter: string[]): Record<string, string> {
+    // TODO: Add a course_id filter if the evaluator is course-specific
+    const filters: Record<string, string> = {};
+
+    if (languageSlugsFilter.length > 0) {
+      filters['language_slugs'] = languageSlugsFilter.join(',');
+    }
+
+    if (courseStageSlugsFilter.length > 0) {
+      filters['course_stage_slugs'] = courseStageSlugsFilter.join(',');
+    }
+
+    return filters;
+  }
+
+  async fetchEvaluations(
+    evaluator: CommunitySolutionEvaluatorModel,
+    course: CourseModel,
+    languageSlugsFilter: string[],
+    courseStageSlugsFilter: string[],
+    resultFilter: string,
+  ): Promise<CommunitySolutionEvaluationModel[]> {
+    const filters = this.buildContextFilters(course, languageSlugsFilter, courseStageSlugsFilter);
+
+    if (resultFilter) {
+      filters['result'] = resultFilter;
+    }
+
+    return (await this.store.query('community-solution-evaluation', {
+      ...filters,
+      ...{
+        evaluator_id: evaluator.id,
+        limit: 30,
+        include: [
+          'community-solution',
+          'community-solution.user',
+          'community-solution.language',
+          'community-solution.course-stage',
+          'community-solution.trusted-evaluations',
+          'community-solution.trusted-evaluations.community-solution',
+          'community-solution.trusted-evaluations.creator',
+          'community-solution.trusted-evaluations.evaluator',
+          'evaluator',
+        ].join(','),
+      },
+    })) as unknown as CommunitySolutionEvaluationModel[];
+  }
+
+  async fetchTrustedEvaluations(
+    evaluator: CommunitySolutionEvaluatorModel,
+    course: CourseModel,
+    languageSlugsFilter: string[],
+    courseStageSlugsFilter: string[],
+  ): Promise<void> {
+    const filters = this.buildContextFilters(course, languageSlugsFilter, courseStageSlugsFilter);
+
+    await this.store.query('trusted-community-solution-evaluation', {
+      ...filters,
+      ...{
+        evaluator_id: evaluator.id,
+        include: [
+          'creator',
+          'community-solution',
+          'community-solution.user',
+          'community-solution.language',
+          'community-solution.course-stage',
+          'community-solution.evaluations',
+          'evaluator',
+        ].join(','),
+      },
+    });
   }
 
   @action
@@ -57,6 +141,11 @@ export default class CodeExampleEvaluatorController extends Controller {
   }
 
   @action
+  async handleDidInsert() {
+    await this.loadEvaluationsTask.perform();
+  }
+
+  @action
   handleRequestedLanguageChange(language: LanguageModel) {
     this.router.transitionTo({ queryParams: { languages: language.slug } });
   }
@@ -71,6 +160,22 @@ export default class CodeExampleEvaluatorController extends Controller {
     });
 
     dummyRecord.unloadRecord();
+  });
+
+  loadEvaluationsTask = task({ keepLatest: true }, async (): Promise<void> => {
+    const { evaluator, course, filteredLanguageSlugs, filteredCourseStageSlugs } = this.model;
+
+    // Load all evaluations in parallel
+    const [passEvaluations, failEvaluations, unsureEvaluations] = await Promise.all([
+      this.fetchEvaluations(evaluator, course, filteredLanguageSlugs, filteredCourseStageSlugs, 'pass'),
+      this.fetchEvaluations(evaluator, course, filteredLanguageSlugs, filteredCourseStageSlugs, 'fail'),
+      this.fetchEvaluations(evaluator, course, filteredLanguageSlugs, filteredCourseStageSlugs, 'unsure'),
+      this.fetchTrustedEvaluations(evaluator, course, filteredLanguageSlugs, filteredCourseStageSlugs),
+    ]);
+
+    this.passEvaluations = passEvaluations;
+    this.failEvaluations = failEvaluations;
+    this.unsureEvaluations = unsureEvaluations;
   });
 
   regenerateAllEvaluationsTask = task({ drop: true }, async (): Promise<void> => {
